Fixed "list index out of range" (issue #2)
This error happens if not even a single category is named. In this case,
the SQL statement in get_categories() yields zero records, causing the error.

To fix this, the categories are now retrieved in two steps: First get
all categories that are used, regardless of whether they're named or
not, and give them dummy names. Then, for the named categories, update
the name with the actual one the user has set. And handle the case where
there are zero named categories.